+++ src/globals/NetworkReplyWatcher.h
@@ -4,6 +4,13 @@
 #include <QObject>
 #include <QTimer>
 
+/**
+ * @brief The NetworkReplyWatcher class takes a QNetworkReply* and watches it.
+ * A timeout is set which aborts the download if no response was received after N seconds.
+ *
+ * @example
+ *   new NetworkReplyWatcher(this, reply) // will delete itself when the reply is deleted
+ */
 class NetworkReplyWatcher : public QObject
 {
     Q_OBJECT
@@ -19,4 +26,6 @@ public slots:
 private:
     QNetworkReply* m_reply;
     QTimer m_timer;
+
+    const int m_timeoutMilliseconds = 3000;
 };