+++ src/tv_shows/SeasonNumber.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <QString>
+#include <ostream>
 
 class SeasonNumber
 {
@@ -23,3 +24,5 @@ public:
 private:
     int m_seasonNumber = -2; // No season; not -1 because Kodi uses it for "no season"
 };
+
+std::ostream& operator<<(std::ostream& os, const SeasonNumber& value);